Create a projection utility where to store projection classes
